+++ js/avatar.js
@@ -1,5 +1,4 @@
-const FILE_TYPES = ['gif', 'jpg', 'jpeg', 'png'];
-const BASE_IMG_CONTENT = '<img src="img/muffin-grey.svg" alt="Аватар пользователя" width="40" height="44">';
+import {BASE_IMG_CONTENT, FILE_TYPES} from './const.js';
 
 const avatarChooser = document.querySelector('#avatar');
 const avatarPreview = document.querySelector('.ad-form-header__preview');
