+++ js/form-activate.js
@@ -1,8 +1,5 @@
 const adForm = document.querySelector('.ad-form');
-
 const mapFilters = document.querySelector('.map__filters');
-
-//неактивное состояние формы
 
 const getDisactiveStateForm = () => {
   adForm.classList.add('ad-form--disabled');
@@ -11,8 +8,6 @@ const getDisactiveStateForm = () => {
   }
 };
 
-// активное состояние формы
-
 const getActiveStateForm = () => {
   adForm.classList.remove('ad-form--disabled');
   for (let i=0; i < adForm.children.length; i++) {
@@ -20,16 +15,12 @@ const getActiveStateForm = () => {
   }
 };
 
-//неактивное состояние фильтров
-
 const getDisactiveStateFilters = () => {
   mapFilters.classList.add('ad-form--disabled');
   for (let i = 0; i < mapFilters.children.length; i++) {
     mapFilters.children[i].disabled = true;
   }
 };
-
-// активное состояние фильтров
 
 const getActiveStateFilters = () => {
   mapFilters.classList.remove('ad-form--disabled');
