+++ js/main.js
@@ -1,4 +1,4 @@
-import './markup.js';
+import './popup.js';
 import './form-valid.js';
 import './map.js';
 import './api.js';
