+++ js/util.js
@@ -1,14 +1,3 @@
-const getRandomInteger = (min, max) => {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  if (min >= max) {
-    throw new Error('Недопустимое значение диапазона');
-  }
-  return Math.floor(Math.random() * (max - min + 1)) + min;};
-
-
-const getRandomArrayElement = (items) => items[getRandomInteger(0, items.length - 1)];
-
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
 const showAlert = (message) => {
@@ -18,7 +7,6 @@ const showAlert = (message) => {
   alertContainer.style.width = '300px';
   alertContainer.style.right = '50%';
   alertContainer.style.transform = 'translateX(50%)';
-  //alertContainer.style.left = 0;
   alertContainer.style.top = '55px';
 
   alertContainer.style.padding = '10px 3px';
@@ -38,44 +26,28 @@ const showAlert = (message) => {
 // Функция debounce - устранения дребезга
 // Источник - https://www.freecodecamp.org/news/javascript-debounce-example
 function debounce (callback, timeoutDelay = 500) {
-  // Используем замыкания, чтобы id таймаута у нас навсегда приклеился
-  // к возвращаемой функции с setTimeout, тогда мы его сможем перезаписывать
   let timeoutId;
 
   return (...rest) => {
-    // Перед каждым новым вызовом удаляем предыдущий таймаут,
-    // чтобы они не накапливались
     clearTimeout(timeoutId);
-
-    // Затем устанавливаем новый таймаут с вызовом колбэка на ту же задержку
     timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
-
-    // Таким образом цикл «поставить таймаут - удалить таймаут» будет выполняться,
-    // пока действие совершается чаще, чем переданная задержка timeoutDelay
   };
 }
 
-// Функция throttle для пропуска кадров:
-// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_throttle
-function throttle (callback, delayBetweenFrames) {
-  // Используем замыкания, чтобы время "последнего кадра" навсегда приклеилось
-  // к возвращаемой функции с условием, тогда мы его сможем перезаписывать
-  let lastTime = 0;
+const getDeclination = (n, form1, form2, form3) => {
+  n = Math.abs(n) % 100;
+  const n1 = n % 10;
+  if (n > 10 && n < 20) {
+    return form3;
+  }
+  if (n1 > 1 && n1 < 5) {
+    return form2;
+  }
+  if (n1 === 1) {
+    return form1;
+  }
+  return form3;
+};
 
-  return (...rest) => {
-    // Получаем текущую дату в миллисекундах,
-    // чтобы можно было в дальнейшем
-    // вычислять разницу между кадрами
-    const now = new Date();
 
-    // Если время между кадрами больше задержки,
-    // вызываем наш колбэк и перезаписываем lastTime
-    // временем "последнего кадра"
-    if (now - lastTime >= delayBetweenFrames) {
-      callback.apply(this, rest);
-      lastTime = now;
-    }
-  };
-}
-
-export {getRandomArrayElement, getRandomInteger, isEscapeKey, showAlert, throttle, debounce};
+export {isEscapeKey, showAlert, debounce, getDeclination};
